+++ app/layouts/auth.vue
