+++ app/pages/dashboard.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script setup>
+definePageMeta({ layout: 'default' })
 
 </script>
